# Faster reservoir sampling