--- a/src/core/space_interaction_API.jl
+++ b/src/core/space_interaction_API.jl
@@ -430,20 +430,21 @@ end
 # Reservoir sampling function (https://en.wikipedia.org/wiki/Reservoir_sampling)
 function resorvoir_sampling_single(iter, model)
     res = iterate(iter)
-    isnothing(res) && return nothing                       # `iterate` returns `nothing` when it ends
+    isnothing(res) && return nothing
     rng = abmrng(model)
+    el, state = res
     w = rand(rng)
     while true
-        choice, state = res                                # random position to return, and the state of the iterator
-        skip_counter = floor(log(rand(rng)) / log(1 - w))  # skip entries in the iterator
+        skip_counter = ceil(Int, randexp(rng)/log(1-w))
         while skip_counter != 0
             skip_res = iterate(iter, state)
-            isnothing(skip_res) && return choice
+            isnothing(skip_res) && return el
             state = skip_res[2]
-            skip_counter -= 1
+            skip_counter += 1
         end
         res = iterate(iter, state)
-        isnothing(res) && return choice
+        isnothing(res) && return el
+        el, state = res
         w *= rand(rng)
     end
 end