This is faster, as some benchmark show, same code and metodology of #788 

before

```julia
julia> @benchmark random_nearby_position($(50,50), $model, $1)
BenchmarkTools.Trial: 10000 samples with 958 evaluations.
 Range (min … max):   94.749 ns …  1.410 μs  ┊ GC (min … max): 0.00% … 90.62%
 Time  (median):     100.439 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   102.630 ns ± 32.078 ns  ┊ GC (mean ± σ):  0.74% ±  2.21%

             ▁▄▇▇██▆▃▁                                          
  ▁▁▁▁▁▁▂▂▃▅▇█████████▇▅▄▃▃▃▃▂▂▂▁▁▁▁▁▁▁▁▁▁▁▂▂▂▂▂▂▂▃▃▂▂▂▂▂▂▂▁▂▁ ▃
  94.7 ns         Histogram: frequency by time          115 ns <

 Memory estimate: 32 bytes, allocs estimate: 1.

julia> @benchmark random_nearby_position($(50,50), $model, $10)
BenchmarkTools.Trial: 10000 samples with 182 evaluations.
 Range (min … max):  594.137 ns … 823.137 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     639.610 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   640.772 ns ±  18.826 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

                 ▃▃▄▆▅▆▆▅▆▇▆█▇▆▆▆▆▅▅▃▃▁▁                         
  ▁▁▁▁▂▂▂▃▃▄▅▇▇███████████████████████████▇▆▅▄▄▃▃▃▂▂▂▂▂▂▂▁▂▁▁▁▁ ▅
  594 ns           Histogram: frequency by time          698 ns <

 Memory estimate: 32 bytes, allocs estimate: 1.
```


after

```julia
julia> @benchmark random_nearby_position($(50,50), $model, $1)
BenchmarkTools.Trial: 10000 samples with 962 evaluations.
 Range (min … max):  89.899 ns …  2.274 μs  ┊ GC (min … max): 0.00% … 93.24%
 Time  (median):     96.220 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   98.950 ns ± 34.391 ns  ┊ GC (mean ± σ):  0.79% ±  2.22%

           ▂▅███▄▁                                             
  ▂▂▁▂▂▂▃▅▇████████▆▅▄▄▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▃▃▃▃▃▃▃▃▃▃▃▃▃▃▃▂▂▂▂▂▂▂ ▃
  89.9 ns         Histogram: frequency by time         118 ns <

 Memory estimate: 32 bytes, allocs estimate: 1.

julia> @benchmark random_nearby_position($(50,50), $model, $10)
BenchmarkTools.Trial: 10000 samples with 359 evaluations.
 Range (min … max):  265.178 ns …  3.143 μs  ┊ GC (min … max): 0.00% … 87.86%
 Time  (median):     286.585 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   288.582 ns ± 40.748 ns  ┊ GC (mean ± σ):  0.19% ±  1.24%

                 ▁▁▃▃▆▇█▇▇▇▇▇▅▆▅▂▃▃▂▃▁▁▁ ▁▁                     
  ▂▁▁▂▂▂▂▂▃▃▃▄▅▆▇█████████████████████████████▆▆▅▅▅▄▄▃▃▃▃▃▃▃▂▃ ▅
  265 ns          Histogram: frequency by time          312 ns <

 Memory estimate: 32 bytes, allocs estimate: 1.
```

This is based on the relationship between an exponential rv and the logarithm of a uniform rv