@Tortar 2023-11-01
> Apparently this PR reduces code coverage by 10% :D nah that's clearly a mistake.

Yes this a bit absurd ahah I noted that already on previous commits, it seems like codecov sometimes randomly doesn't understand what are the files it shouldn't cover



> You mentioned that you wanted to make this into a package. I think it is a great idea. You should, it is definitely helpful for a wide range of applications. But it isn't a priority either!

Working on it at the moment https://github.com/Tortar/IteratorSampling.jl but I'm thinking about some good tricks to make everything better and the package is still not good enough to release 0.1 so I need some more time to finish up the work since uni work is coming for me :D